# docs/reference/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "keepkey-sdk-server",
    "version": "0.1.1",
    "contact": {}
  },
  "servers": [
    {
      "url": "http://localhost:1646/"
    }
  ],
  "paths": {
    "/addresses/utxo": {
      "post": {
        "operationId": "UtxoGetAddress",
        "summary": "Get Binance Address",
        "description": "Get a Bitcoin/Dogecoin/Dash.... cont. (see coin-support) Chain address, loading it into the current session and optionally displaying it on-device.",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "show_display": {
                    "type": "boolean"
                  },
                  "script_type": {
                    "$ref": "#/components/schemas/BTCInputScriptType"
                  },
                  "coin": {
                    "type": "string"
                  },
                  "address_n": {
                    "$ref": "#/components/schemas/types.AddressNList"
                  }
                },
                "required": [
                  "coin",
                  "address_n"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "address": {}
                  },
                  "required": [
                    "address"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Error processing request"
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ],
        "tags": [
          "Address"
        ]
      }
    },
...
  },
  "components": {
...
    "schemas": {
      "types.numeric.UnsignedInteger.Safe": {
        "type": "integer",
        "format": "int64",
        "maximum": 9007199254740991,
        "minimum": 0
      },
      "types.numeric.U32": {
        "format": "uint32",
        "allOf": [
          {
            "$ref": "#/components/schemas/types.numeric.UnsignedInteger.Safe"
          },
          {}
        ],
        "maximum": 4294967295,
        "minimum": 0
      },
      "types.AddressNList": {
        "description": "a BIP-32 path, expressed as a list of node indexes",
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/types.numeric.U32"
        },
        "maxItems": 5,
        "minItems": 3
      },
      "BTCInputScriptType": {
        "type": "string",
        "enum": [
          "cashaddr",
          "bech32",
          "p2pkh",
          "p2sh",
          "external",
          "p2wpkh",
          "p2sh-p2wpkh"
        ]
      },
...
    },
...
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "SESSIONID"
      }
    }
  }
}